Attempting to adjust scripts to be able to work on Ubuntu + have some retries/error reporting

+++ fedora/better-lock-screen.sh
@@ -4,7 +4,12 @@
 USER_HOME=$(getent passwd $SUDO_USER | cut -d: -f6)
 
 cd ${USER_HOME}/src 
-sudo dnf install -y cairo-devel libev-devel libjpeg-devel libjpeg-turbo-devel libxcb-devel libxkbcommon-devel libxkbcommon-x11-devel pam-devel pkg-config xcb-util-devel xcb-util-image-devel
+
+if [ -f /etc/redhat-release ]; then
+  sudo dnf install -y cairo-devel libev-devel libjpeg-devel libjpeg-turbo-devel libxcb-devel libxkbcommon-devel libxkbcommon-x11-devel pam-devel pkg-config xcb-util-devel xcb-util-image-devel
+else
+  sudo apt-get install -y cairo-devel libev-devel libjpeg-devel libjpeg-turbo-devel libxcb-devel libxkbcommon-devel libxkbcommon-x11-devel pam-devel pkg-config xcb-util-devel xcb-util-image-devel
+fi
 
 sudo -u $SUDO_USER bash << EOF
 # i3lock-color is required as a dependency
