+++ fedora/install-optional-software.sh
@@ -20,9 +20,24 @@ sudo dnf install -y flatpack
 flatpak install https://flathub.org/repo/appstream/com.slack.Slack.flatpakref
 
 # Visual Studio Code
-rpm --import https://packages.microsoft.com/keys/microsoft.asc
-sh -c 'echo -e "[code]\nname=Visual Studio Code\nbaseurl=https://packages.microsoft.com/yumrepos/vscode\nenabled=1\ngpgcheck=1\ngpgkey=https://packages.microsoft.com/keys/microsoft.asc" > /etc/yum.repos.d/vscode.repo'
-dnf check-update
-dnf install -y code
+if [ -f /etc/redhat-release ]; then
+  rpm --import https://packages.microsoft.com/keys/microsoft.asc
+  sh -c 'echo -e "[code]\nname=Visual Studio Code\nbaseurl=https://packages.microsoft.com/yumrepos/vscode\nenabled=1\ngpgcheck=1\ngpgkey=https://packages.microsoft.com/keys/microsoft.asc" > /etc/yum.repos.d/vscode.repo'
+  dnf check-update
+  dnf install -y code
+else
+  curl https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > microsoft.gpg
+  install -o root -g root -m 644 microsoft.gpg /etc/apt/trusted.gpg.d/
+  sh -c 'echo "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main" > /etc/apt/sources.list.d/vscode.list'
+  apt-get install apt-transport-https
+  apt-get update
+  apt-get install code # or code-insiders
+fi
 
-dnf install -y redshift
+
+if [ -f /etc/redhat-release ]; then
+  dnf install -y redshift
+else
+  apt-get install redshift
+fi
+
